soc: qcom: rpmh: Make the PM hooks robust and safe to use on RT
The CPU idle driver uses various RPMh hooks in this driver while IRQs are
disabled, which is problematic on RT since normal spin locks are used both
in here and in the generic mailbox API functions.

Since the underlying TCS mailbox driver used by CPU idle is now RT-safe in
atomic context, we can skip the mailbox API functions and go directly to
the TCS mailbox functions in order to bypass the normal spin locks in the
mailbox API. This is fine from a synchronization perspective too, since the
TCS mailbox driver has its own locks to prevent any problematic races.

In addition to this, we can remove the superfluous spin locks used for the
dirty flag and also convert the pool allocator to a fast, lock-less
solution using plain atomics. A fine-grained raw spin lock is now used for
the passthru cache as well.

Finally, the kzalloc() within the sleepable spin-locked section is moved
outside of the lock as an optimization.

Overall, this makes the RPMh driver faster and totally safe on RT.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>