Revert "memlat: Optimize perf event reads when possible"
This reverts commit 190be9ed11703e8e0f7c79461662634792c9beb4.

There are a number of problems with this optimization attempt, such as how
a perf error code could be returned to one of the counter values, a CPU
could enter an SCM call for just one event and result in garbage
computations or a divide-by-zero when calculating the stall percentage,
and IRQs are disabled as a heavy-handed way of preventing CPU migration.

Toss this commit out so that it can be replaced by something better.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: atndko <z1281552865@gmail.com>
(cherry picked from commit 4b1e7a2d84953d8d4fba6b52499000fb86996907)